refactor: move shared sanitizers into base
